# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpServerCreate": {
        "properties": {
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Allowed Tools",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "name": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ],
        "title": "McpServerCreate",
        "type": "object"
      },
      "McpServerList": {
        "properties": {
          "max_servers": {
            "title": "Max Servers",
            "type": "integer"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/McpServerOut"
            },
            "title": "Servers",
            "type": "array"
          }
        },
        "required": [
          "servers",
          "max_servers"
        ],
        "title": "McpServerList",
        "type": "object"
      },
      "McpServerOut": {
        "properties": {
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Tools",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discovered_tools": {
            "items": {
              "$ref": "#/components/schemas/McpToolInfo"
            },
            "title": "Discovered Tools",
            "type": "array"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "header_names": {
            "items": {
              "type": "string"
            },
            "title": "Header Names",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "header_names",
          "allowed_tools",
          "discovered_tools",
          "enabled",
          "last_error",
          "last_checked_at",
          "created_at"
        ],
        "title": "McpServerOut",
        "type": "object"
      },
      "McpServerUpdate": {
        "properties": {
          "allowed_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Tools"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "McpServerUpdate",
        "type": "object"
      },
      "McpToolInfo": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "McpToolInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HootPR API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orgs/{org_slug}/mcp-servers": {
      "get": {
        "operationId": "list_servers_api_orgs__org_slug__mcp_servers_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Servers"
      },
      "post": {
        "operationId": "create_server_api_orgs__org_slug__mcp_servers_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Server"
      }
    },
    "/api/orgs/{org_slug}/mcp-servers/{server_id}": {
      "delete": {
        "operationId": "delete_server_api_orgs__org_slug__mcp_servers__server_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Server"
      },
      "patch": {
        "operationId": "update_server_api_orgs__org_slug__mcp_servers__server_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Server"
      }
    },
    "/api/orgs/{org_slug}/mcp-servers/{server_id}/discover": {
      "post": {
        "description": "Connect to the server (from the API process, SSRF-guarded) and cache its tool list.",
        "operationId": "discover_tools_api_orgs__org_slug__mcp_servers__server_id__discover_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Tools"
      }
    },
...
  }
}